clang-tidy: fixed the issues found by the widening, special member and other hand fixed checks
- bugprone-implicit-widening-of-multiplication-result: the multiplications whose result is
  widened (buffer sizes, strides and pointer offsets) are now performed in the wide type by
  casting an operand (std::size_t, std::ptrdiff_t or VkDeviceSize) instead of the result.
- cppcoreguidelines-special-member-functions:
  - classes that delete or declare their copy and move members also declare the destructor
  - the scoped state helpers delete the copy operations
  - classes whose declared special members were all defaulted now follow the rule of zero
- bugprone-unused-local-non-trivial-variable: removed two unused strings.
- bugprone-switch-missing-default-case: added the default case to the EGL attribute switch.
- bugprone-derived-method-shadowing-base-method: removed WindowInputEvent::SYS_DoConstruct, which
  only forwarded to the identical WindowTransactionEvent method.
- bugprone-empty-catch: the UIReader try parse fallbacks are marked with NOLINT.
- portability-template-virtual-member-function: VulkanNativeWindowAdapter is marked with NOLINT
  as MSVC instantiates all virtual members, so the Windows build verifies them.

+++ DemoApps/GLES2/Blur/source/GausianHelper.cpp
@@ -136,7 +136,7 @@ namespace Fsl
   void GausianHelper::CalculateGausianKernel(std::vector<double>& rKernel, const int32_t length, const double sigma)
   {
     assert((length & 1) == 1);
-    rKernel.resize(length * length);
+    rKernel.resize(static_cast<std::size_t>(length) * length);
 
     const double calculatedEuler = 1.0 / (2.0 * MathHelper::PI * sigma * sigma);
