+++ DemoApps/GLES3/E7_0_ParticleSystem/source/E7_0_ParticleSystem.cpp
@@ -91,7 +91,7 @@ namespace Fsl
 
     for (int i = 0; i < NUM_PARTICLES; i++)
     {
-      float* particleData = &m_userData.particleData[i * PARTICLE_SIZE];
+      float* particleData = &m_userData.particleData[static_cast<std::ptrdiff_t>(i) * PARTICLE_SIZE];
 
       // Lifetime of particle
       (*particleData++) = (static_cast<float>(rand() % 10000) / 10000.0f);
