+++ DemoApps/GLES3/FurShellRendering/source/Shader/FurShaderBase.cpp
@@ -265,8 +265,8 @@ namespace Fsl
       m_locLightDirection[0] = GL_CHECK(glGetUniformLocation(hProgram, "LightDirection"));
       m_locLightColor[0] = GL_CHECK(glGetUniformLocation(hProgram, "LightColor"));
       m_locLightCount = GL_CHECK(glGetUniformLocation(hProgram, "LightCount"));
-      m_lightDirection.resize(3 * lightCount);
-      m_lightColor.resize(3 * lightCount);
+      m_lightDirection.resize(static_cast<std::size_t>(3) * lightCount);
+      m_lightColor.resize(static_cast<std::size_t>(3) * lightCount);
       std::fill(m_lightDirection.begin(), m_lightDirection.end(), 0.0f);
       std::fill(m_lightColor.begin(), m_lightColor.end(), 0.0f);
     }
