+++ DemoApps/GLES3/ParticleSystem/source/PS/Draw/ParticleDrawQuadsGLES3.cpp
@@ -141,7 +141,7 @@ namespace Fsl
     // FSLLOG3_INFO("Particles-ToVertices Time: " << end - start);
 
     glBindBuffer(m_pCurrentBuffer->GetTarget(), m_pCurrentBuffer->Get());
-    m_pCurrentBuffer->SetDataFast(0, m_buffer.data(), particleCount * 6);
+    m_pCurrentBuffer->SetDataFast(0, m_buffer.data(), static_cast<std::size_t>(particleCount) * 6);
     m_pCurrentBuffer->EnableAttribArrays(m_particleAttribLink);
 
     glDrawArrays(GL_TRIANGLES, 0, UncheckedNumericCast<GLsizei>(particleCount * 6));
