+++ DemoFramework/FslGraphics/include/FslGraphics/Bitmap/ReadOnlyRawBitmap.hpp
@@ -129,7 +129,7 @@ namespace Fsl
                                               const BitmapOrigin origin)
     {
       const auto stride = PixelFormatUtil::CalcMinimumStride(sizePx.Width(), pixelFormat);
-      if ((sizePx.RawUnsignedHeight() * stride) > span.size())
+      if ((static_cast<std::size_t>(sizePx.RawUnsignedHeight()) * stride) > span.size())
       {
         throw std::invalid_argument("span is not large enough to contain a bitmap of the given dimensions");
       }
@@ -143,7 +143,7 @@ namespace Fsl
       {
         throw std::invalid_argument("stride is smaller than the width allows");
       }
-      if ((sizePx.RawUnsignedHeight() * stride) > span.size())
+      if ((static_cast<std::size_t>(sizePx.RawUnsignedHeight()) * stride) > span.size())
       {
         throw std::invalid_argument("span is not large enough to contain a bitmap of the given dimensions");
       }
